Agrega alta medico

+++ TPC_Brandan_Repolledo/Dominio/Especialidad.cs
@@ -8,7 +8,7 @@ namespace Dominio
 {
     public class Especialidad
     {
-        public int IdEspecialidad { get; set; }
+        public long IdEspecialidad { get; set; }
 
         public string Nombre { get; set; }
